Added comments to KrunchCVWidget and new code.
Added a ton of comments explaining in detail how the vision processing
works along with a function that bounds the azimuth to positive and
negative 180 degrees. The update to the class also enables sending of
the azimuth, range, and the rpms of the shooter wheels to the robot in
real time.