Got SmartDashboard to work again with new...
... camera tilt variables (high and low). Also found out that
smartdashboard needs to be run as admin if there is no CSV file in order
to generate a new, default one (otherwise a CSV error will appear).
Another bug found is that the VPLayout.xml will only save when
processImage is running, so only save when an image is on the screen
otherwise the entire VPLayout.xml will be erased of its contents.