Initial commit for KrunchCV tester. This is a...
...derivation of the KrunchCV widget that does vision processing. You
run the jar in a terminal with the arguments of filepaths to different
images to process. It appears with the raw and processed image. It
communicates over Network Tables, requires a KrunchCVSettings.txt in a
C:\Program Files\SmartDashboard, and is not a widget. Everything else
about it is the same.